+++ src/index.ts
@@ -10,4 +10,6 @@ export {
 export { Propagation } from './Propagation'
 export { IsolationLevel } from './IsolationLevel'
 export { Transactional } from './Transactional'
+export { runInTransaction } from './runInTransaction';
+export { wrapInTransaction } from './wrapInTransaction';
 export * from './TransactionalError'
